requirements.txt
asn1crypto==0.23.0
astroid==1.5.3
attrs==17.2.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
isort==4.2.15
Jinja2==2.9.6
lazy-object-proxy==1.3.1
livereload==2.5.1
lxml==4.0.0
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
mkdocs==0.16.3
parsel==1.2.0
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
pylint==1.7.4
pymongo==3.5.1
pyOpenSSL==17.3.0
PyYAML==3.12
queuelib==1.4.2
requests==2.18.4
ruamel.yaml==0.15.34
Scrapy==1.4.0
service-identity==17.0.0
six==1.11.0
tornado==4.5.2
Twisted==17.9.0
urllib3==1.22
w3lib==1.18.0
wrapt==1.10.11
zope.interface==4.4.3
